ENH: 'closed' option in numpy.ma.extras._ezclump
Adds a flag, 'closed', to increase the range stop by 1.
For compatibility, the default value is False.

The motivation for this feature comes from clustering together elements
of an array by timestamp.
I have a signal from a detector with an unevenly spaced time series,
from which an event is defined by a cluster of entries that occur less
than X time after the previous entry.
As such, with `numpy.ma.diff` and a filter `x[x <= max_time]` I get a
mask with the entries that have an entry shortly after. However, this
does not include the last entry that is close to the previous one, and
should be included in the cluster, because it is not close to the entry
after it.
Currently, this can be solved by adding 1 to the end of each slice (which
can be done at the time of using the slice).

It is useful in this usecase, and possibly others, as it saves the
necessity of iterating over the array to add 1 to the end of each slice.

The term is consistent with pandas rolling.